requirements.txt
Django==5.0.7
django-debug-toolbar==4.4.5
django-jazzmin==3.0.0
djangorestframework==3.15.2
ipython==8.26.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1